iot_Project/requirements.txt
pyserial
adafruit-io
keras
tensorflow
Pillow
numpy
opencv-python
flask